refactor(detector): move detector to banks namespace
