+++ src/monopoly/banks/__init__.py
@@ -1,16 +1,17 @@
 import logging
 from typing import Type
 
-from ..examples.example_bank import ExampleBank
 from .base import BankBase
 from .citibank import Citibank
 from .dbs import Dbs
+from .detector import BankDetector
+from .example_bank import ExampleBank
 from .hsbc import Hsbc
 from .maybank import Maybank
 from .ocbc import Ocbc
 from .standard_chartered import StandardChartered
 
-banks: list[Type[BankBase]] = [
+banks: list[Type["BankBase"]] = [
     Citibank,
     Dbs,
     ExampleBank,
@@ -21,3 +22,5 @@ banks: list[Type[BankBase]] = [
 ]
 
 logger = logging.getLogger(__name__)
+
+__all__ = ["BankDetector", "BankBase", *banks]
