--- src/monopoly/examples/__init__.py
@@ -1,3 +1,0 @@
-from .example_bank import ExampleBank
-
-__all__ = ["ExampleBank"]
